Move library to action_class and rename properties
Get things ready to add a :create action. This makes the properties generic enough that it would work.
